+++ entrypoints/admin/src/main.tsx
@@ -5,6 +5,7 @@ import zhCN from "antd/es/locale/zh_CN";
 import 'dayjs/locale/zh-cn';
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
+import '@ant-design/v5-patch-for-react-19';
 import 'uno.css';
 import App from "./App.tsx";
 import "./index.css";